feat: expose programmatic api
This refactors the main convertToPdf function and exposes it from `util/md-to-pdf.js`, so it can be shared between the new programmatic API and the CLI. The CLI has been moved to `cli.js` and `index.js` exposes a wrapper around `mdToPdf` instead.

Closes #25
